Move api folder to shared
