+++ src/__tests__/client/api/sendYmlData.test.ts
@@ -1,5 +1,5 @@
 import axios from "axios";
-import startScenario from "client/api/sendYmlData";
+import startScenario from "src/client/shared/api/sendYmlData";
 import { YmlRequestType } from "src/client/yml/entities/form";
 
 jest.mock("axios");
