+++ src/__tests__/client/api/userRequests.test.ts
@@ -1,5 +1,5 @@
 import axios from "axios";
-import { checkToken, loginUserByAdminPanel } from "client/api/userRequests";
+import { checkToken, loginUserByAdminPanel } from "src/client/shared/api/userRequests";
 import { AuthByAdminPanelRequestBody } from "src/declarations";
 
 jest.mock("axios");
