+++ src/client/shared/api/scenarioRequests.ts
