+++ src/client/shared/api/sendYmlData.ts
@@ -1,5 +1,5 @@
 import axios from "axios";
-import { YmlRequestType } from "../yml/entities/form";
+import { YmlRequestType } from "../../yml/entities/form";
 
 const sendYmlData = async (data: YmlRequestType) =>
   await axios.post("/api/yml/start", data, {
